Added possibility to calculate recipe details, removed coupling between RecipeDetails/Recipe/Ingredient

+++ UnitTests/Ingredients/Entity/IngredientTests.cs
@@ -64,9 +64,9 @@ namespace WhatShouldIEat.AdministrationService.Tests.Ingredients.Entity
 			var result = _systemUnderTest.GetMacroNutrientQuantity(100);
 			using (new AssertionScope())
 			{
-				result[MacroNutrient.Carbohydrate].Should().Be(20);
-				result[MacroNutrient.Fat].Should().Be(20);
-				result[MacroNutrient.Protein].Should().Be(30);
+				result[(int)MacroNutrient.Carbohydrate].Should().Be(20);
+				result[(int)MacroNutrient.Fat].Should().Be(20);
+				result[(int)MacroNutrient.Protein].Should().Be(30);
 			}
 		}
